api/requirements.txt
Flask==2.3.2
requests==2.25.1
gunicorn==20.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Migrate==4.0.4
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.5
pycryptodome==3.18.0
PyJWT==2.7.0
Werkzeug==2.3.6
